refactor(rust): isolate xlang benchmarks
