+++ benchmarks/rust/README.md
@@ -1,12 +1,20 @@
 # Fory Rust Benchmark
 
-This benchmark compares Rust serialization and deserialization throughput for Apache Fory and Protocol Buffers using the shared benchmark dataset defined in `benchmarks/proto/bench.proto`.
+This directory orchestrates two independent Rust benchmark packages:
+
+- `xlang/` compares Apache Fory, Protocol Buffers, and MessagePack with the
+  shared models defined by `benchmarks/proto/bench.proto`.
+- `local/` owns Rust-local buffer and threading benchmarks plus
+  external-type serialization comparisons.
+
+Neither package depends on the other, so building a local benchmark does not
+compile xlang models or serializer monomorphizations.
 
 ## Prerequisites
 
-The benchmark build generates Rust code from the shared schema and requires a
+The xlang package generates Rust code from the shared schema and requires a
 `protoc` executable on `PATH`, or the `PROTOC` environment variable set to an
-existing executable.
+existing executable. The local package does not require `protoc`.
 
 ## Quick Start
 
@@ -25,10 +33,11 @@ cd benchmarks/rust
 Options:
   --data <struct|sample|mediacontent|structlist|samplelist|mediacontentlist>
                                Filter benchmark by data type
-  --serializer <fory|protobuf>
+  --serializer <fory|protobuf|msgpack>
                                Filter benchmark by serializer
   --filter <regex>             Custom criterion filter
   --no-report                  Skip Python report generation
+  --no-copy-docs               Skip copying the report into docs/benchmarks/rust
 ```
 
 Examples:
@@ -64,9 +73,10 @@ where one int32 field is widened to int64.
 | `SampleList`        | List of shared `Sample` payloads                                       |
 | `MediaContentList`  | List of shared `MediaContent` payloads                                 |
 
-The separate `external_type_bench` Criterion target contains branch-local Rust
-external-type serialization comparisons. Keeping it separate preserves the
-ordinary `serialization_bench` binary and measurement shape. Each comparison has `self_serialize`,
+The local package's separate `external_type_bench` Criterion target contains
+branch-local Rust external-type serialization comparisons. Its package
+boundary preserves the xlang `serialization_bench` binary, dependency graph,
+and measurement shape. Each comparison has `self_serialize`,
 `selected_serialize`, `self_deserialize`, and `selected_deserialize` lanes.
 The self lane uses an equivalent self-provided Rust target. The selected lane
 uses the external structural serializer, manual serializer, carrier serializer,
@@ -90,9 +100,9 @@ The matrix covers:
 Run one comparison case by its Criterion group name:
 
 ```bash
-cargo bench --bench external_type_bench -- carrier_map_nested
-cargo bench --bench external_type_bench -- external_command_compatible
-cargo bench --bench external_type_bench -- dynamic_trait_arc
+cargo bench --manifest-path local/Cargo.toml --bench external_type_bench -- carrier_map_nested
+cargo bench --manifest-path local/Cargo.toml --bench external_type_bench -- external_command_compatible
+cargo bench --manifest-path local/Cargo.toml --bench external_type_bench -- dynamic_trait_arc
 ```
 
 For regression gates, run baseline and current cases sequentially. Existing
@@ -102,7 +112,9 @@ lane; do not add compatibility shims to the baseline checkout.
 
 ## Shared Proto Schema
 
-The Rust benchmark uses the shared protobuf definition at `benchmarks/proto/bench.proto`, the same benchmark schema used by the C++ benchmark suite.
+The Rust xlang package uses the shared protobuf definition at
+`benchmarks/proto/bench.proto`, the same benchmark schema used by the C++
+benchmark suite.
 
 ## Manual Commands
 
@@ -110,24 +122,32 @@ Run Criterion benchmarks:
 
 ```bash
 cd benchmarks/rust
-cargo bench --bench serialization_bench
-cargo bench --bench external_type_bench
+cargo bench --manifest-path xlang/Cargo.toml --bench serialization_bench
+cargo bench --manifest-path local/Cargo.toml --bench external_type_bench
 ```
 
 Print serialized sizes:
 
 ```bash
 cd benchmarks/rust
-cargo run --release --bin fory_profiler -- --print-all-serialized-sizes
+cargo run --release --manifest-path xlang/Cargo.toml --bin fory_profiler -- --print-all-serialized-sizes
+```
+
+Build either package independently:
+
+```bash
+cd benchmarks/rust
+cargo check --manifest-path xlang/Cargo.toml --all-targets
+cargo check --manifest-path local/Cargo.toml --all-targets
 ```
 
 Generate the markdown report manually:
 
 ```bash
 cd benchmarks/rust
-cargo bench --bench serialization_bench 2>&1 | tee results/cargo_bench.log
-cargo bench --bench external_type_bench 2>&1 | tee -a results/cargo_bench.log
-cargo run --release --bin fory_profiler -- --print-all-serialized-sizes | tee results/serialized_sizes.txt
+cargo bench --manifest-path xlang/Cargo.toml --bench serialization_bench 2>&1 | tee results/cargo_bench.log
+cargo bench --manifest-path local/Cargo.toml --bench external_type_bench 2>&1 | tee -a results/cargo_bench.log
+cargo run --release --manifest-path xlang/Cargo.toml --bin fory_profiler -- --print-all-serialized-sizes | tee results/serialized_sizes.txt
 python benchmark_report.py --log-file results/cargo_bench.log --size-file results/serialized_sizes.txt --output-dir results
 ```
