+++ benchmarks/rust/local/Cargo.toml
@@ -15,10 +15,34 @@
 # specific language governing permissions and limitations
 # under the License.
 
-[workspace]
-members = ["local", "xlang"]
-resolver = "2"
+[package]
+name = "fory-rust-local-benchmarks"
+version.workspace = true
+edition.workspace = true
 
-[workspace.package]
-version = "1.5.0-alpha.0"
-edition = "2021"
+[[bench]]
+name = "external_type_bench"
+path = "benches/external_type_bench.rs"
+harness = false
+
+[[bench]]
+name = "buffer_write_bench"
+path = "benches/buffer_write_bench.rs"
+harness = false
+
+[[bench]]
+name = "buffer_read_bench"
+path = "benches/buffer_read_bench.rs"
+harness = false
+
+[[bench]]
+name = "threaded_bench"
+path = "benches/threaded_bench.rs"
+harness = false
+
+[dependencies]
+criterion = "0.5"
+fory = { path = "../../../rust/fory" }
+fory-core = { path = "../../../rust/fory-core" }
+fory-derive = { path = "../../../rust/fory-derive" }
+fory-external-model = { path = "../../../rust/api-tests/external-model" }
