+++ benchmarks/rust/local/benches/buffer_write_bench.rs
