+++ benchmarks/rust/local/benches/external_type_bench.rs
