+++ benchmarks/rust/xlang/Cargo.toml
@@ -15,10 +15,28 @@
 # specific language governing permissions and limitations
 # under the License.
 
-[workspace]
-members = ["local", "xlang"]
-resolver = "2"
+[package]
+name = "fory-rust-xlang-benchmarks"
+version.workspace = true
+edition.workspace = true
 
-[workspace.package]
-version = "1.5.0-alpha.0"
-edition = "2021"
+[[bin]]
+name = "fory_profiler"
+path = "src/main.rs"
+
+[[bench]]
+name = "serialization_bench"
+path = "benches/serialization_bench.rs"
+harness = false
+
+[dependencies]
+clap = { version = "4.0", features = ["derive"] }
+criterion = "0.5"
+fory = { path = "../../../rust/fory" }
+fory-derive = { path = "../../../rust/fory-derive" }
+prost = "0.12"
+rmp-serde = "1.3"
+serde = { version = "1.0", features = ["derive"] }
+
+[build-dependencies]
+prost-build = "0.12"
