+++ benchmarks/rust/xlang/src/lib.rs
