+++ benchmarks/rust/xlang/src/serializers/protobuf.rs
